verlet fix, anda el main

+++ src/inicializar.h
@@ -1,7 +1,7 @@
 #ifndef INICIALIZAR_H
 #define INICIALIZAR_H
 int Inicializar (double* vector_posvel,double* vector_fuerza,int N,double* LUTF,int Ntable,double rho, double m, double T);
-int Llenar_Pos (double* vector,int N,double rho);
+double Llenar_Pos (double* vector,int N,double rho);
 int Llenar_Vel (double* vector,int N,double rho,double m, double T);
 double Rand_Gauss (double m, double T);
